Fix build failure when enable-go-utils is empty
The current Makefile was not importing the top-level Makefile.global.
The Makefile.global is included in order to read configuration variables
in submake files. We want to adopt this pattern as it makes the Makefile
easy to reason about for reading the output of configure.
This commit fixes that and changes the logic of the Makefile in go-utils
to be truthy.

Signed-off-by: Nadeem Ghani <nghani@pivotal.io>