gp_upgrade: prepare shutdown-cluster
Trying to set some "patterns" for testing goroutines. The idea is to
test the portion outside the goroutine, with the dependency for the
goroutine call injected, so you can stub it out totally inert. Then you
can unit test the class that was injected for the function call that
happens in the goroutine.

New structs and interfaces for upgradeConfig, "ClusterPair" and the
logger to separate them out

Tried to make logging thread safe. We're using channels to pass
information around and have main method write to the file.

Initiate a pattern for knowing the state of asynchronous goroutines.
 - Leave files behind about the initiation of action, progress,
 finished-ness
 - TODO: work on `status` calls which would read these files

Author: Marbin Tan <mtan@pivotal.io>
Author: Nadeem Ghani <nghani@pivotal.io>
Author: C.J. Jameson <cjameson@pivotal.io>